+++ bittytax/conv/parsers/__init__.py
@@ -4,6 +4,7 @@ from . import barclays
 from . import bitfinex
 from . import bitstamp
 from . import bittrex
+from . import bittylicious
 from . import blockchain
 from . import blockfi
 from . import bscscan